pyproject.toml
[tool.poetry]
name = "species"
version = "0.1.0"
description = ""
authors = ["umschaudhary <umesschaudhary@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
autopep8 = "^1.6.0"
uvicorn = "^0.17.6"
ipdb = "^0.13.9"
pydantic = {extras = ["email"], version = "^1.9.1"}
alembic = "^1.8.0"
SQLAlchemy = "^1.4.37"
python-dotenv = "^0.20.0"
asyncpg = "^0.25.0"
Jinja2 = "^3.1.2"
fastapi-utils = "^0.2.1"
psycopg2-binary = "^2.9.3"
passlib = "^1.7.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
